refactor: 네이밍 컨벤션 및 app 상태 co-locate
