+++ src/entities/storage/model/storage.ts
@@ -1,4 +1,4 @@
-import { useWebExtensionStorageLocal } from './useWebExtensionStorage';
+import { useWebExtensionStorageLocal } from './use-web-extension-storage';
 
 export const useStorageDemo = () => {
   return useWebExtensionStorageLocal('webext-demo', 'Storage Demo');
