+++ src/entities/storage/model/use-web-extension-storage.ts
