+++ src/popup/Popup.tsx
@@ -4,6 +4,10 @@ import Logo from '~/components/Logo';
 import { useStorageDemo } from '~/entities/storage';
 import { Timer } from '~/features/timer';
 
+const openOptionsPage = () => {
+  browser.runtime.openOptionsPage();
+};
+
 const Popup = () => {
   const [storageDemo] = useStorageDemo();
 
@@ -12,7 +16,7 @@ const Popup = () => {
       <Logo />
       <div>Popup</div>
       <p className="mt-2 opacity-50">This is the popup page</p>
-      <button className="btn mt-2" onClick={browser.runtime.openOptionsPage}>
+      <button className="btn mt-2" onClick={openOptionsPage}>
         Open Options
       </button>
       <div className="mt-2">
